votings_backend/requirements.txt
asgiref==3.2.3
Django==3.0.2
django-cors-headers==3.2.1
django-extensions==2.2.6
djangorestframework==3.11.0
pytz==2019.3
six==1.14.0
sqlparse==0.3.0
